+++ packages/client/src/game/gameState.ts
@@ -18,4 +18,6 @@ export const updateFrameTime = (ts: number) => {
 
 export const getScreenScale = () => min(gl.drawingBufferWidth, gl.drawingBufferHeight) / GAME_CFG._camera._size;
 
-export const getNameByClientId = (client: ClientID) => client === clientId ? clientName : remoteClients.get(client)?._name;
+export function getNameByClientId(client: ClientID) {
+    return client === clientId ? clientName : remoteClients.get(client)?._name;
+}