+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add simple autoplay mode for debugging
 - Change user settings data
+- Add replay files initial support (wip)
 
 ## 1.0.13